move ResourceType into fiat-api
