+++ fiat-core/src/main/java/com/netflix/spinnaker/fiat/model/resources/Resource.java
@@ -18,6 +18,7 @@ package com.netflix.spinnaker.fiat.model.resources;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.netflix.spinnaker.fiat.api.Permissions;
+import com.netflix.spinnaker.fiat.api.ResourceType;
 
 public interface Resource {
   String getName();
